Switch to Font Awesome and update Instagram icon, take 2
The previous icon pack has not been maintained so we moved to
Font-Awesome. The goal here is to make the switch to font-awesome as
backwards compatible as possible with existing Mito sites. Important
note: the footer's background color must be defined, even when the
background is image-based. Removed the recently added compass color
utility, since access to the contrast-color sass function is no longer
useful due to backwards compatibility concerns.

https://app.asana.com/0/164883087213679/164671092599667